+++ container/container.h
@@ -13,26 +13,60 @@ extern "C" {
 /*----------------------------------------------------------------------------
                       C O N T A I N E R _ V R _ T
 ----------------------------------------------------------------------------*/
+
 typedef struct {
-	fmu_vr_t			fmu_vr;
-	long				fmu_id;
+	fmu_vr_t					fmu_vr;
+	unsigned long				fmu_id;
 } container_vr_t;
 
 
 /*----------------------------------------------------------------------------
                       C O N T A I N E R _ P O R T _ T
 ----------------------------------------------------------------------------*/
+
 typedef struct {
     unsigned long               nb;	 /* number of connected FMU from a container port */
     container_vr_t              *links;
 } container_port_t;
 
 
+/*----------------------------------------------------------------------------
+                       C O N T A I N E R _ C L O C K _ T
+----------------------------------------------------------------------------*/
+
+typedef struct {
+	unsigned long				fmu_id;
+	unsigned long				nb;
+	fmu_vr_t					*fmu_vr;
+	unsigned long				*local_clock_vr;
+} container_clock_t;
+
+
+/*----------------------------------------------------------------------------
+                C O N T A I N E R _ C L O C K _ L I S T _ T
+----------------------------------------------------------------------------*/
+
+typedef struct {
+	unsigned long				nb_fmu;
+	unsigned long				nb_clocks;
+	container_clock_t			*clocks;
+	double						next_event_time;
+	int							nb_active_clocks;
+	unsigned long				*next_active_clocks;
+} container_clock_list_t;
+
+
+/*----------------------------------------------------------------------------
+            C O N T A I N E R _ D O _ S T E P _ F U N C T I O N _ T
+----------------------------------------------------------------------------*/
+
 typedef fmu_status_t (*container_do_step_function_t)(struct container_s *container);
+
 
 /*----------------------------------------------------------------------------
                             C O N T A I N E R _ T
 ----------------------------------------------------------------------------*/
+
 typedef struct container_s {
 	/* configuration */
 	int							profiling;
@@ -59,6 +93,8 @@ typedef struct container_s {
 	DECLARE_LOCAL(booleans, int);
 	DECLARE_LOCAL(booleans1, bool);
 	DECLARE_LOCAL(strings, char *);
+	DECLARE_LOCAL(binaries, fmu_binary_t);
+	DECLARE_LOCAL(clocks, bool);
 #undef DECLARE_LOCAL
 
 	/* container ports definition */
@@ -80,6 +116,8 @@ typedef struct container_s {
     DECLARE_PORT(booleans);
     DECLARE_PORT(booleans1);
     DECLARE_PORT(strings);
+	DECLARE_PORT(binaries);
+	DECLARE_PORT(clocks);
 #undef DECLARE_PORT
 
 	convert_table_t				conversions;
@@ -93,11 +131,11 @@ typedef struct container_s {
 	double						stop_time;
 	int							tolerance_defined;
 	int							stop_time_defined;
+	container_clock_list_t		local_clocks;
 
 	fmi2CallbackAllocateMemory	allocate_memory;		/* used to embed FMU-2.0 */
 	fmi2CallbackFreeMemory      free_memory;			/* used to embed FMU-2.0 */
 } container_t;
-
 
 
 /*----------------------------------------------------------------------------